Do away with stdint.h
- The types used for sizes (size_t, ssize_t) are now 32 bits to fully
  comply with the ansi standard
- Make rf_Assert take in a "proper" bool instead of a typedefed uintptr

+++ src/rf_AsciiString.c
@@ -3,8 +3,8 @@
 
 struct rf_AsciiString {
 	char *chars;
-	i64 str_len;
-	i64 capacity;
+	isize str_len;
+	isize capacity;
 };
 
 b32 rf_AsciiString_test(void);
